requirements.txt

# === Core ===
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# === IPC ===
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# === LLM Provider ===
requests>=2.31.0
tenacity>=8.2.0

# === Voice ===
# TTS (Legacy pins for compatibility)
numpy<2.0.0
scipy<1.12.0
espnet==202301
git+https://github.com/robinhad/ukrainian-tts.git

# STT
faster-whisper>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1

# === Memory ===
chromadb>=0.4.22
redis>=5.0.0

# === Vision & Automation ===
pyautogui>=0.9.54
pillow>=10.0.0
pyobjc-framework-Quartz>=10.0 ; sys_platform == 'darwin'
pyobjc-framework-ApplicationServices>=10.0 ; sys_platform == 'darwin'

# === Utils ===
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
mcp>=1.0.0
# fastmcp removed: use mcp.server.FastMCP from the 'mcp' package instead of external fastmcp

pyobjc >= 10.0 ; sys_platform == 'darwin'
psutil>=5.9.0

# === Database ===
sqlalchemy>=2.0.0
asyncpg>=0.29.0
greenlet>=3.0.0